+++ lib/models/definitions/transaction/take_profit_details.ex
@@ -13,7 +13,7 @@ defmodule ExOanda.TakeProfitDetails do
 
   typed_embedded_schema do
     field(:price, :float)
-    field(:time_in_force, Ecto.Enum, values: ~w(GTC GTD GFD FOK IOC)a)
+    field(:time_in_force, Ecto.Enum, values: ~w(GTC GTD GFD FOK IOC)a, default: :GTC)
     field(:gtd_time, :utc_datetime_usec)
     embeds_one :client_extensions, ClientExtensions
   end